ai/main.py
import os
from fastapi import FastAPI, Depends, HTTPException
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
import httpx

app = FastAPI()

oauth2_scheme = OAuth2PasswordBearer(tokenUrl='token')

# Security configuration
@app.get('/secure')
def secure_route(token: str = Depends(oauth2_scheme)):
    if not validate_token(token):
        raise HTTPException(status_code=401, detail='Invalid or expired token')
    return {'message': 'Secure Endpoint Accessed'}

# SAP AI Core integration
class PredictionRequest(BaseModel):
    input_data: dict

@app.post('/predict')
def predict(request: PredictionRequest, token: str = Depends(oauth2_scheme)):
    if not validate_token(token):
        raise HTTPException(status_code=401, detail='Invalid or expired token')
    response = call_sap_ai_core(request.input_data)
    return response

# Helper functions

def validate_token(token: str) -> bool:
    # Placeholder function for token validation logic
    return token == 'valid-token'

def call_sap_ai_core(input_data: dict) -> dict:
    # Placeholder function for calling SAP AI Core services
    url = 'https://sap-ai-core-endpoint.com/predict'
    headers = {'Authorization': f'Bearer {os.getenv('SAP_AI_CORE_API_KEY')}', 'Content-Type': 'application/'}
    response = httpx.post(url, headers=headers, ={'input': input_data})
    return response.()
